Create Box Calculator
This is a calculator that has three types of boxes, a large box holds five items, a medium box holds three items, and a large box holds one item. It sorts through the given amount of items and places as many into large boxes until there is less than five items, then if there are three or four items remaining it will put them in a medium box and the remaining items after that are put into small boxes.